[MM-42739] Move insights to api4/insights and keep time range as string until required
