+++ api4/api.go
@@ -134,6 +134,9 @@ type Routes struct {
 	SharedChannels *mux.Router // 'api/v4/sharedchannels'
 
 	Permissions *mux.Router // 'api/v4/permissions'
+
+	InsightsForTeam *mux.Router // 'api/v4/teams/{team_id:[A-Za-z0-9]+}/top'
+	InsightsForUser *mux.Router // 'api/v4/users/me/top'
 }
 
 type API struct {
@@ -255,6 +258,9 @@ func Init(srv *app.Server) (*API, error) {
 
 	api.BaseRoutes.Permissions = api.BaseRoutes.APIRoot.PathPrefix("/permissions").Subrouter()
 
+	api.BaseRoutes.InsightsForTeam = api.BaseRoutes.Team.PathPrefix("/top").Subrouter()
+	api.BaseRoutes.InsightsForUser = api.BaseRoutes.Users.PathPrefix("/me/top").Subrouter()
+
 	api.InitUser()
 	api.InitBot()
 	api.InitTeam()
@@ -296,6 +302,7 @@ func Init(srv *app.Server) (*API, error) {
 	api.InitSharedChannels()
 	api.InitPermissions()
 	api.InitExport()
+	api.InitInsights()
 	if err := api.InitGraphQL(); err != nil {
 		return nil, err
 	}
