+++ api4/channel_test.go
@@ -4549,172 +4549,3 @@ func TestViewChannelWithoutCollapsedThreads(t *testing.T) {
 	require.NoError(t, err)
 	require.Zero(t, threads.TotalUnreadMentions)
 }
-
-func TestGetTopChannelsForTeamSince(t *testing.T) {
-	th := Setup(t).InitBasic()
-	defer th.TearDown()
-
-	th.ConfigStore.SetReadOnlyFF(false)
-	defer th.ConfigStore.SetReadOnlyFF(true)
-	th.App.UpdateConfig(func(cfg *model.Config) { cfg.FeatureFlags.InsightsEnabled = true })
-
-	client := th.Client
-	userId := th.BasicUser.Id
-
-	channel4 := th.CreatePublicChannel()
-	channel5 := th.CreatePrivateChannel()
-	channel6 := th.CreatePrivateChannel()
-	th.App.AddUserToChannel(th.BasicUser, channel4, false)
-	th.App.AddUserToChannel(th.BasicUser, channel5, false)
-	th.App.AddUserToChannel(th.BasicUser, channel6, false)
-
-	channelIDs := [6]string{th.BasicChannel.Id, th.BasicChannel2.Id, th.BasicPrivateChannel.Id, channel4.Id, channel5.Id, channel6.Id}
-
-	i := len(channelIDs)
-	for _, channelID := range channelIDs {
-		for j := i; j > 0; j-- {
-			_, _, err := client.CreatePost(&model.Post{UserId: userId, ChannelId: channelID, Message: "zz" + model.NewId() + "a"})
-			require.NoError(t, err)
-		}
-		i--
-	}
-
-	teamId := th.BasicChannel.TeamId
-
-	expectedTopChannels := []struct {
-		ID           string
-		MessageCount int64
-	}{
-		{ID: th.BasicChannel.Id, MessageCount: 7},
-		{ID: th.BasicChannel2.Id, MessageCount: 5},
-		{ID: th.BasicPrivateChannel.Id, MessageCount: 4},
-		{ID: channel4.Id, MessageCount: 3},
-		{ID: channel5.Id, MessageCount: 2},
-	}
-
-	t.Run("get-top-channels-for-team-since", func(t *testing.T) {
-		topChannels, _, err := client.GetTopChannelsForTeamSince(teamId, model.TimeRangeToday, 0, 5)
-		require.NoError(t, err)
-
-		for i, channel := range topChannels.Items {
-			assert.Equal(t, expectedTopChannels[i].ID, channel.ID)
-			assert.Equal(t, expectedTopChannels[i].MessageCount, channel.MessageCount)
-		}
-
-		topChannels, _, err = client.GetTopChannelsForTeamSince(teamId, model.TimeRangeToday, 1, 5)
-		require.NoError(t, err)
-		assert.Equal(t, channel6.Id, topChannels.Items[0].ID)
-		assert.Equal(t, int64(1), topChannels.Items[0].MessageCount)
-	})
-
-	t.Run("get-top-channels-for-user-since exclude channels user is not member of", func(t *testing.T) {
-		excludedChannel := th.CreatePrivateChannel()
-
-		for i := 0; i < 10; i++ {
-			_, _, err := client.CreatePost(&model.Post{UserId: userId, ChannelId: excludedChannel.Id, Message: "zz" + model.NewId() + "a"})
-			require.NoError(t, err)
-		}
-
-		th.RemoveUserFromChannel(th.BasicUser, excludedChannel)
-
-		topChannels, _, err := client.GetTopChannelsForTeamSince(teamId, model.TimeRangeToday, 0, 5)
-		require.NoError(t, err)
-
-		for i, channel := range topChannels.Items {
-			assert.Equal(t, expectedTopChannels[i].ID, channel.ID)
-			assert.Equal(t, expectedTopChannels[i].MessageCount, channel.MessageCount)
-		}
-	})
-
-	t.Run("get-top-channels-for-team-since invalid team id", func(t *testing.T) {
-		_, resp, err := client.GetTopChannelsForTeamSince("12345", model.TimeRangeToday, 0, 5)
-		assert.Error(t, err)
-		CheckBadRequestStatus(t, resp)
-
-		_, resp, err = client.GetTopChannelsForTeamSince(model.NewId(), model.TimeRangeToday, 0, 5)
-		assert.Error(t, err)
-		CheckNotFoundStatus(t, resp)
-	})
-
-	t.Run("get-top-channels-for-team-since not a member of team", func(t *testing.T) {
-		th.UnlinkUserFromTeam(th.BasicUser, th.BasicTeam)
-		_, resp, err := client.GetTopChannelsForTeamSince(teamId, model.TimeRangeToday, 0, 5)
-		assert.Error(t, err)
-		CheckForbiddenStatus(t, resp)
-	})
-}
-
-func TestGetTopChannelsForUserSince(t *testing.T) {
-	th := Setup(t).InitBasic()
-	defer th.TearDown()
-
-	th.ConfigStore.SetReadOnlyFF(false)
-	defer th.ConfigStore.SetReadOnlyFF(true)
-	th.App.UpdateConfig(func(cfg *model.Config) { cfg.FeatureFlags.InsightsEnabled = true })
-
-	client := th.Client
-	userId := th.BasicUser.Id
-
-	channel4 := th.CreatePublicChannel()
-	channel5 := th.CreatePrivateChannel()
-	channel6 := th.CreatePrivateChannel()
-	th.App.AddUserToChannel(th.BasicUser, channel4, false)
-	th.App.AddUserToChannel(th.BasicUser, channel5, false)
-	th.App.AddUserToChannel(th.BasicUser, channel6, false)
-
-	channelIDs := [6]string{th.BasicChannel.Id, th.BasicChannel2.Id, th.BasicPrivateChannel.Id, channel4.Id, channel5.Id, channel6.Id}
-
-	i := len(channelIDs)
-	for _, channelID := range channelIDs {
-		for j := i; j > 0; j-- {
-			_, _, err := client.CreatePost(&model.Post{UserId: userId, ChannelId: channelID, Message: "zz" + model.NewId() + "a"})
-			require.NoError(t, err)
-		}
-		i--
-	}
-
-	teamId := th.BasicChannel.TeamId
-
-	expectedTopChannels := []struct {
-		ID           string
-		MessageCount int64
-	}{
-		{ID: th.BasicChannel.Id, MessageCount: 6},
-		{ID: th.BasicChannel2.Id, MessageCount: 5},
-		{ID: th.BasicPrivateChannel.Id, MessageCount: 4},
-		{ID: channel4.Id, MessageCount: 3},
-		{ID: channel5.Id, MessageCount: 2},
-	}
-
-	t.Run("get-top-channels-for-user-since", func(t *testing.T) {
-		topChannels, _, err := client.GetTopChannelsForUserSince(teamId, model.TimeRangeToday, 0, 5)
-		require.NoError(t, err)
-
-		for i, channel := range topChannels.Items {
-			assert.Equal(t, expectedTopChannels[i].ID, channel.ID)
-			assert.Equal(t, expectedTopChannels[i].MessageCount, channel.MessageCount)
-		}
-
-		topChannels, _, err = client.GetTopChannelsForUserSince("", model.TimeRangeToday, 1, 5)
-		require.NoError(t, err)
-		assert.Equal(t, channel6.Id, topChannels.Items[0].ID)
-		assert.Equal(t, int64(1), topChannels.Items[0].MessageCount)
-	})
-
-	t.Run("get-top-channels-for-user-since invalid team id", func(t *testing.T) {
-		_, resp, err := client.GetTopChannelsForUserSince("12345", model.TimeRangeToday, 0, 5)
-		assert.Error(t, err)
-		CheckBadRequestStatus(t, resp)
-
-		_, resp, err = client.GetTopChannelsForUserSince(model.NewId(), model.TimeRangeToday, 0, 5)
-		assert.Error(t, err)
-		CheckNotFoundStatus(t, resp)
-	})
-
-	t.Run("get-top-channels-for-user-since not a member of team", func(t *testing.T) {
-		th.UnlinkUserFromTeam(th.BasicUser, th.BasicTeam)
-		_, resp, err := client.GetTopChannelsForUserSince(teamId, model.TimeRangeToday, 0, 5)
-		assert.Error(t, err)
-		CheckForbiddenStatus(t, resp)
-	})
-}
