+++ api4/reaction.go
@@ -16,9 +16,6 @@ func (api *API) InitReaction() {
 	api.BaseRoutes.Post.Handle("/reactions", api.APISessionRequired(getReactions)).Methods("GET")
 	api.BaseRoutes.ReactionByNameForPostForUser.Handle("", api.APISessionRequired(deleteReaction)).Methods("DELETE")
 	api.BaseRoutes.Posts.Handle("/ids/reactions", api.APISessionRequired(getBulkReactions)).Methods("POST")
-
-	api.BaseRoutes.Team.Handle("/top/reactions", api.APISessionRequired(getTopReactionsForTeamSince)).Methods("GET")
-	api.BaseRoutes.Users.Handle("/me/top/reactions", api.APISessionRequired(getTopReactionsForUserSince)).Methods("GET")
 }
 
 func saveReaction(c *Context, w http.ResponseWriter, r *http.Request) {
@@ -139,87 +136,5 @@ func getBulkReactions(c *Context, w http.ResponseWriter, r *http.Request) {
 		c.Err = model.NewAppError("getBulkReactions", "api.marshal_error", nil, jsonErr.Error(), http.StatusInternalServerError)
 		return
 	}
-	w.Write(js)
-}
-
-func getTopReactionsForTeamSince(c *Context, w http.ResponseWriter, r *http.Request) {
-	c.RequireTeamId().RequireTimeRange()
-	if c.Err != nil {
-		return
-	}
-
-	team, err := c.App.GetTeam(c.Params.TeamId)
-	if err != nil {
-		c.Err = err
-		return
-	}
-
-	if (!team.AllowOpenInvite || team.Type != model.TeamOpen) && !c.App.SessionHasPermissionToTeam(*c.AppContext.Session(), team.Id, model.PermissionViewTeam) {
-		c.SetPermissionError(model.PermissionViewTeam)
-		return
-	}
-
-	topReactionList, err := c.App.GetTopReactionsForTeamSince(c.Params.TeamId, c.AppContext.Session().UserId, &model.InsightsOpts{
-		StartUnixMilli: c.Params.TimeRange,
-		Page:           c.Params.Page,
-		PerPage:        c.Params.PerPage,
-	})
-	if err != nil {
-		c.Err = err
-		return
-	}
-
-	js, jsonErr := json.Marshal(topReactionList)
-	if jsonErr != nil {
-		c.Err = model.NewAppError("getTopReactionsForTeamSince", "api.marshal_error", nil, jsonErr.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	w.Write(js)
-}
-
-func getTopReactionsForUserSince(c *Context, w http.ResponseWriter, r *http.Request) {
-	c.RequireTimeRange()
-	if c.Err != nil {
-		return
-	}
-
-	c.Params.TeamId = r.URL.Query().Get("team_id")
-
-	// TeamId is an optional parameter
-	if c.Params.TeamId != "" {
-		if !model.IsValidId(c.Params.TeamId) {
-			c.SetInvalidURLParam("team_id")
-			return
-		}
-
-		team, teamErr := c.App.GetTeam(c.Params.TeamId)
-		if teamErr != nil {
-			c.Err = teamErr
-			return
-		}
-
-		if (!team.AllowOpenInvite || team.Type != model.TeamOpen) && !c.App.SessionHasPermissionToTeam(*c.AppContext.Session(), team.Id, model.PermissionViewTeam) {
-			c.SetPermissionError(model.PermissionViewTeam)
-			return
-		}
-	}
-
-	topReactionList, err := c.App.GetTopReactionsForUserSince(c.AppContext.Session().UserId, c.Params.TeamId, &model.InsightsOpts{
-		StartUnixMilli: c.Params.TimeRange,
-		Page:           c.Params.Page,
-		PerPage:        c.Params.PerPage,
-	})
-	if err != nil {
-		c.Err = err
-		return
-	}
-
-	js, jsonErr := json.Marshal(topReactionList)
-	if jsonErr != nil {
-		c.Err = model.NewAppError("getTopReactionsForUserSince", "api.marshal_error", nil, jsonErr.Error(), http.StatusInternalServerError)
-		return
-	}
-
 	w.Write(js)
 }
